+++ packages/cli/src/init.ts
@@ -1,7 +1,7 @@
 import { CommandModule } from 'yargs';
 import { main } from '@angular-devkit/schematics-cli/bin/schematics';
 export const initCommand: CommandModule = {
-    command: ['init'],
+    command: ['init', '$0'],
     describe: 'Init config',
     builder: yargs => {
         yargs.option('mode', {