+++ packages/template/src/services/global-context.spec.ts
@@ -1,10 +1,10 @@
 import { DocgeniSiteConfig } from './../interfaces/config';
 import { CONFIG_TOKEN, GlobalContext } from './global-context';
 import { createServiceFactory, createHttpFactory, SpectatorHttp, HttpMethod } from '@ngneat/spectator';
-import { NavigationItem } from '../interfaces';
+import { DocItem, NavigationItem } from '../interfaces';
 import { TestBed } from '@angular/core/testing';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { HttpClient } from '@angular/common/http';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
+import { HttpClient, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { Location } from '@angular/common';
 
 describe('GlobalContext', () => {
@@ -34,7 +34,8 @@ describe('GlobalContext', () => {
 
     beforeEach(() => {
         TestBed.configureTestingModule({
-            imports: [HttpClientTestingModule]
+            imports: [],
+            providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
         });
     });
 
@@ -206,7 +207,7 @@ describe('GlobalContext', () => {
                 items: [{ id: '3', title: '', path: '' }]
             },
             { id: '', title: '', path: '', items: [{ id: '4', title: '', path: '', hidden: true }] }
-        ];
+        ] as NavigationItem[];
         const globalContext = createGlobalContext({} as DocgeniSiteConfig);
         const result = globalContext.sortDocItems(list);
         expect(result.length).toBe(3);