+++ packages/toolkit/src/fs/docgeni-scoped.ts
@@ -5,14 +5,6 @@ export class DocgeniScopedHost<T extends object> extends virtualFs.ResolverHost<
         super(delegate);
     }
 
-    // protected _resolve(path: Path): Path {
-    //     if (nodePath.isAbsolute(path)) {
-    //         return path;
-    //     } else {
-    //         return nodePath.join(this.root, path) as Path;
-    //     }
-    // }
-
     protected _resolve(path: Path): Path {
         if (isAbsolute(path)) {
             return path;