+++ scripts/sync-template-version.ts
@@ -1,6 +1,6 @@
 import { toolkit } from '@docgeni/toolkit';
 import * as path from 'path';
-import * as yargs from 'yargs-parser';
+import yargs from 'yargs-parser';
 import writeJsonFile from 'write-json-file';
 
 async function sync() {
@@ -13,7 +13,7 @@ async function sync() {
         // corePackageJson.peerDependencies['@docgeni/template'] = toVersion;
         await writeJsonFile(corePackageJsonPath, corePackageJson, {
             detectIndent: true,
-            indent: 2
+            indent: 2,
         });
         toolkit.print.info(`sync core/package.json 's @docgeni/template to ${toolkit.print.chalk.green(toVersion)}`);
     }